pyproject.toml
[project]
name = "pilflow"
version = "0.1.0"
requires-python = ">=3.8"
description = "A Pillow-based image processing pipeline project."
authors = [
    { name = "Cheng Yanru", email = "yanru@cyanru.com" },
]
dependencies = [
    "Pillow>=6.2.1",
    "jinnang>=0.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "requests>=2.28.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"